Remove loading screen
There is a massive amount of users using this css, and they all have a loading screen whenever you visit their profile. However, for anyone wanting to visit the profile or even just going through random, this is very annoying. As most users import this css directly, this would resolve this issue.